Fix absurd_extreme_comparisons clippy error in check_avoid_edges
Wholesale-copying ztworldmgr.rs from the original detour-testing-framework
tip reintroduces the pre-existing bug where avoid_edges - 1 stays
unsigned, making radius < 0 always false and tripping clippy's
deny-by-default absurd_extreme_comparisons lint. Cast to i32 before
subtracting, same fix already applied on main.

+++ openzt/src/ztworldmgr.rs
@@ -254,7 +254,7 @@ impl BFEntity {
     }
 
     pub fn check_avoid_edges(&self, tile: &BFTile) -> bool {
-        let radius = self.entity_type().avoid_edges - 1;
+        let radius = self.entity_type().avoid_edges as i32 - 1;
 
         if radius < 0 {
             return false;
